Clean up mobile styling. Choose layout based on device.
