+++ app.vue
@@ -1,6 +1,14 @@
 <template>
   <NuxtRouteAnnouncer />
-  <NuxtLayout>
+  <NuxtLayout :name="isMobile ? 'mobile-flux' : 'default'">
     <NuxtPage />
   </NuxtLayout>
 </template>
+
+<script setup>
+const { isMobile } = useDevice()
+
+definePageMeta({
+  layout: false
+})
+</script>
