+++ nuxt.config.ts
@@ -4,9 +4,10 @@ export default defineNuxtConfig({
   devtools: { enabled: true },
   modules: [
     '@nuxt/ui',
-    '@pinia/nuxt',
     '@nuxt/image',
+    '@pinia/nuxt',
     'nuxt-tiptap-editor',
+    '@nuxtjs/device',
   ],
   tiptap: {
     prefix: 'Tiptap', //prefix for Tiptap imports, composables not included
